fix: simplified adc init, removed redundant comments and Serial prints

+++ AcousticsMicrocontrollerCode/teensy41/src/ethernet_module.cpp
@@ -70,7 +70,7 @@ void UDP_send_ready_signal(uint8_t* remoteIPArray, uint16_t remotePort) {
 
     // Start sending data
     Udp.beginPacket(remoteIP, remotePort);
-    for (int16_t i = 0; i < 5; i++) {
+    for (size_t i = 0; i < 5; i++) {
         tempByte = (byte)UDPReplyBuffer[i];
         Udp.write(tempByte);
     }
